requirements.txt
# THESE ARE TENTATIVE AND MAYBE BE CHANGED FOR CONFIRMED VERSIONS INDICATE WITH A TICK ✅

Flask==3.0.3
Flask-CORS==4.0.0                       # For handling Cross-Origin Resource Sharing
langchain==0.2.13                   # Updated to match langchain-community requirement
langchain-community==0.2.12
langchain-core==0.2.30
google-generativeai==0.8.3          # For Gemini 1.5 API
sentence-transformers==3.0.1        # For All-MiniLM-L6-v2 embeddings and cross-encoder reranking
psycopg2-binary==2.9.9              # PostgreSQL adapter
pgvector==0.3.4                     # pgvector support for SQLAlchemy
sqlalchemy==2.0.30                  # ORM for database interactions
python-dotenv==1.0.1                # For managing environment variables (e.g., API keys)
boto3==1.34.124                     # AWS SDK (for S3 file handling if done in-app)
unstructured==0.15.7                # For robust document parsing (PDF, DOCX, PPTX, etc.)
pdf2image==1.17.0                   # Optional: for PDF rendering if needed
Pillow==10.3.0                      # Required by pdf2image
transformers==4.41.2                # Underlying library for sentence-transformers (includes cross-encoders)
torch==2.3.1                        # PyTorch (required by sentence-transformers & transformers)